fix bug logic of get invigilator report

+++ src/main/java/com/fullsnacke/eimsfuhcmbe/dto/response/InvigilatorAssignmentReportResponseDTO.java
@@ -17,7 +17,7 @@ public class InvigilatorAssignmentReportResponseDTO {
     int totalRequiredInvigilationSlots;
     int totalNonInvigilatedSlots;
 
-    int totalAssignedHours;
+    double totalAssignedHours;
     double totalInvigilatedHours;
     double totalRequiredInvigilationHours;
